+++ CHANGES.rst
@@ -1,6 +1,13 @@
 astropy-helpers Changelog
 *************************
 
+2.0.5 (unreleased)
+------------------
+
+- Fix segmentation faults that occurred when the astropy-helpers submodule
+  was first initialized in packages that also contained Cython code. [#375]
+
+
 2.0.4 (2018-02-09)
 ------------------